Fix for int64 and float64 truncation
Bug: 157123549
Bug: 156000505
Test: ./gradlew :sqlite:sqlite-inspection:cC

+++ sqlite/sqlite-inspection/src/main/java/androidx/sqlite/inspection/SqliteInspector.java
@@ -615,10 +615,10 @@ final class SqliteInspector extends Inspector {
                 builder.setStringValue(cursor.getString(index));
                 break;
             case Cursor.FIELD_TYPE_INTEGER:
-                builder.setIntValue(cursor.getInt(index));
+                builder.setLongValue(cursor.getLong(index));
                 break;
             case Cursor.FIELD_TYPE_FLOAT:
-                builder.setFloatValue(cursor.getFloat(index));
+                builder.setDoubleValue(cursor.getDouble(index));
                 break;
         }
